Prueba/requirements_prueba.txt
attrs==23.1.0
azure-core==1.28.0
azure-storage-blob==12.17.0
azure-storage-file-datalake==12.12.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
idna==3.4
isodate==0.6.1
jsonlines==3.1.0
pycparser==2.21
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
